create health_check route
Created the /api/health_check/:room_name route to check the number
of players inside a room and if the room is online. If the room is
available, the server returns a status 200, with the number of players
inside it and max number of players supported by the room.

If the room is unavailable the server returns a status 404 with players
and max players equal to 0

Also added the prefix /api to every custom server route