fix a security issue in authentication flow
Added an await statement before comparePassword function calls.
This function is async, and it returns a Promise, the Promise was being
used as the criteria for authenticating the user on login and other
operations, so it was always "truthy". This was causing issues like
an user being authenticated even with a wrong password!